Move PersonPropertyEditor to persons package
